+++ src/main/resources/io/jenkins/plugins/finitestatethirdpartyupload/docker/Dockerfile
