+++ src/main/resources/io/jenkins/plugins/finitestatethirdpartyupload/docker/test_upload_test_results.py
