+++ src/main/resources/io/jenkins/plugins/finitestatethirdpartyupload/docker/test_utils.py
