+++ src/main/resources/io/jenkins/plugins/finitestatethirdpartyupload/docker/upload_test_results.py
